eliminacion de archivos duplicados por archivos con sweetalert
