+++ src/components/Alert/Alert.jsx
@@ -1,6 +1,5 @@
 import {useEffect } from 'react';
 import Swal from 'sweetalert2';
-
 import withReactContent from "sweetalert2-react-content";
 import './Alert.css';
