+++ src/components/Banner/Banner.jsx
@@ -5,15 +5,21 @@ import CarruselHorizontal from "../Carrusel/CarruselHorizontal/CarruselHorizonta
 import Boton from '../Boton/Boton';
 import Loader from '../Loader/Loader';
 import Alert from '../Alert/Alert';
-import Swal from "sweetalert2";
 import { db } from "../../firebase/Firebase";
 import { addDoc, getDocs, collection, query, where } from "firebase/firestore";
 import { useAuth } from '../../context/authContext.js';
+import colors from '../../config/config.js'
 import "./Banner.css";
 
 const IMAGE_PATH = process.env.REACT_APP_URL_IMAGE_TMDB;
 
-export const TrailerPlayer = ({ trailer, closeBanner }) => {
+export const TrailerPlayer = ({ trailer, closeTrailer, onTrailerFinish }) => {
+    const handleTrailerEnd = () => {
+        if (onTrailerFinish) {
+        onTrailerFinish();
+        }
+    };
+
     return (
         <div className="youtube-container">
             <YouTube
@@ -31,21 +37,142 @@ export const TrailerPlayer = ({ trailer, closeBanner }) => {
                 showinfo: 0,
             },
         }}
+        onEnd={handleTrailerEnd}  // Añade este manejo de evento
         />
-        <Boton texto={'close'} funcion={closeBanner}/>
+        <Boton width={'100px'} texto={'close'} funcion={closeTrailer} backgroundColor={colors.azul} backgroundHover={colors.naranja}/>
     </div>
     );
 };
 
-const BannerContent = ({ movie, trailer, setPlaying, closeBanner }) => {
-
+const BannerContent = ({ movie, trailer, setPlaying, closeBanner, handleRecent, handleList, showAlert, setShowAlert, alertConfig, setAlertConfig}) => {
     const [playButtonClicked, setPlayButtonClicked] = useState(false);
-    const { user } = useAuth();
-    const [userEmail, setUserEmail] = useState(null);
 
     const handlePlayButtonClick = () => {
         setPlaying(true);
         setPlayButtonClicked(true);
+    };
+
+    useEffect(() => {
+        if (playButtonClicked) {
+            handleRecent();
+        }
+    }, [playButtonClicked]);
+
+    useEffect(() => {
+        // Mostrar la alerta cuando alertConfig se establezca
+        if (showAlert && alertConfig) {
+            setAlertConfig({
+                title: alertConfig.title,
+                text: alertConfig.text,
+                icon: alertConfig.icon,
+                confirmButtonText: alertConfig.confirmButtonText,
+                showCancelButton: alertConfig.showCancelButton,
+                onConfirm: alertConfig.onConfirm,
+                onCancel: alertConfig.onCancel,
+            });
+
+            // Restablecer el estado de showAlert y alertConfig
+            setShowAlert(false);
+        }
+    }, [showAlert, alertConfig]);
+
+
+    return (
+        <div className="container-banner">
+            <div className="banner">
+                <div className="botones-banner">
+                    {/* Botón de Reproducir Trailer */}
+                    {trailer ? (
+                        <Boton  texto={'Play Trailer'} 
+                                funcion={handlePlayButtonClick} 
+                                width={'10rem'} 
+                                fontSize={'1rem'}
+                                margin={'0.5rem'}
+                                height={'3rem'}
+                                backgroundColor={colors.naranja}
+                                backgroundHover={colors.azul}/>
+                    ) : (
+                        <Boton  texto={'Play Trailer'} 
+                                funcion={() => setPlayButtonClicked(true)} 
+                                width={'10rem'} 
+                                fontSize={'1rem'}
+                                margin={'0.5rem'}
+                                height={'3rem'}
+                                backgroundColor = {colors.naranja}
+                                backgroundHover={colors.azul}/>
+                    )}
+                    {/* Botón de Agregar a Mi Lista */}
+                    <Boton  texto={'Agregar a mi lista'} 
+                                funcion={handleList} 
+                                width={'10rem'} 
+                                fontSize={'1rem'}
+                                margin={'0.5rem'}
+                                height={'3rem'}
+                                backgroundColor = {colors.naranja}
+                                backgroundHover={colors.azul}/>
+                    {/* Botón de Volver al Listado */}
+                    <Boton  texto={'Volver al Listado'} 
+                                funcion={closeBanner} 
+                                width={'10rem'} 
+                                fontSize={'1rem'}
+                                margin={'0.5rem'}
+                                height={'3rem'}
+                                backgroundColor = {colors.naranja}
+                                backgroundHover={colors.azul}/>
+                    {/* Alerta para No Hay Trailer */}
+                    {playButtonClicked && !trailer && (
+                        <Alert
+                            title={"Sin trailer"}
+                            text={"Lo sentimos, el trailer no está disponible"}
+                            icon={"info"}
+                            confirmButtonText={"OK"}
+                            showCancelButton={false}
+                            onCancel={() => setPlayButtonClicked(false)} // Restablecer el estado de playButtonClicked al cancelar
+                        />
+                    )}
+                    {/* Alerta para Agregar a Lista o Recientes */}
+                    {showAlert && (
+                        <Alert
+                            title={alertConfig.title}
+                            text={alertConfig.text}
+                            icon={alertConfig.icon}
+                            confirmButtonText={alertConfig.confirmButtonText}
+                            showCancelButton={alertConfig.showCancelButton}
+                        />
+                    )}
+                </div>
+            <h1 className="titulo-banner">{movie.title}</h1>
+            <p className="descripcion-banner">{movie.overview}</p>
+        </div>
+    </div>
+    );
+};
+
+function Banner() {
+    const [showCardContainer, setShowCardContainer] = useState(true);
+    const [selectedMovie, setSelectedMovie] = useState(null);
+    const [movies, setMovies] = useState([]);
+    const [trailer, setTrailer] = useState(null);
+    const [actualPage, setActualPage] = useState(0);
+    const [movie, setMovie] = useState({ title: "Loading Movies" });
+    const [playing, setPlaying] = useState(false);
+    const [isLoadingMovies, setIsLoadingMovies] = useState(true);
+
+    const [showAlert, setShowAlert] = useState(false);  // Nuevo estado para controlar la visibilidad del Alert
+    const [alertConfig, setAlertConfig] = useState(null);
+    const { user } = useAuth();
+    const [userEmail, setUserEmail] = useState(null);
+
+    // Función para manejar la finalización del trailer
+    const handleTrailerFinish = () => {
+        setShowAlert(true);
+        setAlertConfig({
+            icon: "success",
+            title: "Agregado a recientes",
+            text: "El vídeo se ha agregado a recientes.",
+            confirmButtonText: "OK",
+            showCancelButton: false,
+        });
         handleRecent();
     };
 
@@ -54,19 +181,24 @@ const BannerContent = ({ movie, trailer, setPlaying, closeBanner }) => {
         if (querySnapshot.size === 0) {
             // No hay documentos con el mismo ID: agregarlo
             await addDoc(route, dataToAdd);
-            Swal.fire({
-                icon: "success",
-                title: "Agregado correctamente",
-                text: "La película o serie se ha agregado.",
-            });
-            //console.log('Agregado correctamente');
+            setShowAlert(true);  // Mostrar el Alert después de agregar correctamente
+            setAlertConfig({
+                icon:"success",
+                title:"Agregado correctamente",
+                text:"La película o serie se ha agregado.",
+                confirmButtonText:"OK",
+                showCancelButton:false,
+            }) 
         } else {
             // Ya existe un documento con el mismo ID
-            Swal.fire({
-                icon: "info",
-                title: "Id repetido",
-                text: "Lo siento, La película o serie ya está guardada en la sección correspondiente.",
-            });
+            setShowAlert(true);  // Mostrar el Alert en caso de ID repetido
+            setAlertConfig({
+                icon:"info",
+                title:"Id repetido",
+                text:"Lo siento, La película o serie ya está guardada en la sección correspondiente.",
+                confirmButtonText:"OK",
+                showCancelButton:false,
+            })
         }
     }
 
@@ -103,63 +235,9 @@ const BannerContent = ({ movie, trailer, setPlaying, closeBanner }) => {
         if (movie.original_title) {
             dataToAdd2.nombre = movie.original_title;
             await handleListOrRecent(route2, dataToAdd2, await getDocs(query(route2, where("id", "==", movie.id))));
+            
         }
     }
-
-
-    return (
-        <div className="container-banner">
-            <div className="banner">
-                <div className="botones-banner">
-                    {trailer ? (
-                        <button
-                            className="boton-banner"
-                            onClick={handlePlayButtonClick}
-                            type="button">
-                            Play Trailer
-                        </button>
-                    ) : (
-                        <button
-                            className="boton-banner"
-                            onClick={() => setPlayButtonClicked(true)} // Simula el clic para mostrar el Alert
-                            type="button">
-                            Play Trailer
-                        </button>
-                    )}
-                    <button
-                        className="boton-banner"
-                        onClick={() => handleList()}
-                        type="button">
-                        Agregar a mi lista
-                    </button>
-
-                    {playButtonClicked && !trailer && (
-                        <Alert
-                            title={'Sin trailer'}
-                            text={'Lo sentimos, el trailer no está disponible'}
-                            icon={'info'}
-                            confirmButtonText={'Aceptar'}
-                            cancelButtonText={'Cancelar'}
-                        />
-                    )}
-                </div>
-            <h1 className="titulo-banner">{movie.title}</h1>
-            <p className="descripcion-banner">{movie.overview}</p>
-        </div>
-    </div>
-    );
-};
-
-function Banner() {
-    const [showCardContainer, setShowCardContainer] = useState(true);
-    const [selectedMovie, setSelectedMovie] = useState(null);
-    const [movies, setMovies] = useState([]);
-    const [trailer, setTrailer] = useState(null);
-    const [actualPage, setActualPage] = useState(0);
-    const [movie, setMovie] = useState({ title: "Loading Movies" });
-    const [playing, setPlaying] = useState(false);
-    const [isLoadingMovies, setIsLoadingMovies] = useState(true);
-
 
     const fetchMovies = async () => {
         try {
@@ -199,9 +277,24 @@ function Banner() {
     };
 
     const closeBanner = () => {
-        setPlaying(false);
+        //setPlaying(false);
         setShowCardContainer(true);
+       // setAlertConfig(null);  // Restablecer alertConfig cuando se cierra el banner
     };
+
+    const closeTrailer = ( ) => {
+        setPlaying(false);
+    }
+
+    // Efecto secundario para restablecer alertConfig cuando se desmonta el componente o cambia showCardContainer
+    useEffect(() => {
+        return () => {
+            if (alertConfig !== null) {
+                setAlertConfig(null);
+            }
+        };
+        
+    }, [showCardContainer]);
 
     useEffect(() => {
         fetchMovies();
@@ -220,9 +313,24 @@ function Banner() {
                     style={{
                     backgroundImage: `url("${IMAGE_PATH}${movie.backdrop_path}")`}}>
                     {playing ? (
-                    <TrailerPlayer trailer={trailer} closeBanner={closeBanner} />
+                    <TrailerPlayer 
+                        trailer={trailer} 
+                        closeTrailer={closeTrailer} 
+                        onTrailerFinish={handleTrailerFinish} 
+                    />
                     ) : (
-                    <BannerContent movie={movie} trailer={trailer} setPlaying={setPlaying} closeBanner={closeBanner} />
+                    <BannerContent 
+                        movie={movie} 
+                        trailer={trailer} 
+                        setPlaying={setPlaying} 
+                        closeBanner={closeBanner}
+                        handleRecent={handleRecent}
+                        handleList={handleList}
+                        showAlert={showAlert}
+                        setShowAlert={setShowAlert}
+                        alertConfig={alertConfig}
+                        setAlertConfig={setAlertConfig}
+                    />                      
                     )}
                 </main>
             )}
