+++ src/components/IntroVideo/IntroVideo.css
@@ -1,6 +1,6 @@
-#introVideo {
-  width: 100%;
-  height: 100vh;
+.video-js {
+  width: 100vw;
+  height: calc(100vh);
   /* 100% del viewport height */
   object-fit: cover;
   /* Para asegurar que el video cubra completamente el contenedor */
@@ -9,8 +9,8 @@
 
 .intro-video-container {
   position: relative;
-  top: 0;
-  left: 0;
+  /* top: 0;
+  left: 0; */
   width: 100%;
   height: 100%;
   z-index: 1000;
@@ -23,21 +23,7 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  background: #003686;
-  border-radius: 30px;
-  color: #fff;
-  padding: 10px 20px;
-  cursor: pointer;
-  border: none;
-  outline: none;
   z-index: 1100;
-  font-weight: normal;
-    font-size: 20px;
-    font-family: 'Roboto';
-}
-.play-intro-button:hover{
-  color:white;
-  background: var(--color-naranja);
 }
 
 /* Estilos para el botón "Omitir Intro" */
@@ -45,25 +31,8 @@
   position: absolute;
   top: 10px;
   right: 10px;
-  background: #003686;
-  border-radius: 30px;
-  color: #fff;
-  padding: 10px 20px;
-  cursor: pointer;
-  border: none;
-  outline: none;
   z-index: 1100;
-  font-weight: normal;
-  font-size: 20px;
-  font-family: 'Roboto';
 }
-.skip-intro-button:hover {
-  color:white;
-  background: var(--color-naranja);
-  transform: scale(1.05);
-
-}
-
 /* IntroVideo.css */
 
 /* #introVideo::-webkit-media-controls {
