--- src/components/IntroVideo/IntroVideo2.css
@@ -1,45 +1,0 @@
-.video-js {
-  width: 100vw;
-  height: calc(100vh);
-  /* 100% del viewport height */
-  object-fit: cover;
-  /* Para asegurar que el video cubra completamente el contenedor */
-  outline: none;
-}
-
-.intro-video-container {
-  position: relative;
-  /* top: 0;
-  left: 0; */
-  width: 100%;
-  height: 100%;
-  z-index: 1000;
-  /* Asegúrate de que el video esté en la parte superior de la pila de capas */
-}
-
-/* Estilos para el botón "Play Intro" */
-.play-intro-button {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: 1100;
-}
-
-/* Estilos para el botón "Omitir Intro" */
-.skip-intro-button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  z-index: 1100;
-}
-/* IntroVideo.css */
-
-/* #introVideo::-webkit-media-controls {
-  display: none !important;
-} */
-
-/* #introVideo {
-  outline: none;
-  Evita el contorno predeterminado en algunos navegadores al enfocar 
-} */
