+++ src/components/RenderItems/RenderItemsContainer/RenderItemsContainer.jsx
@@ -8,8 +8,7 @@ import { db } from '../../../firebase/Firebase.js';
 import { deleteDoc, doc } from 'firebase/firestore';
 import { useNavigate  } from 'react-router-dom';
 import Loader from '../../Loader/Loader';
-import Swal from "sweetalert2";
-import withReactContent from "sweetalert2-react-content";
+import Alert from '../../Alert/Alert'
 
 const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
     const [movies, setMovies] = useState([]);
@@ -18,11 +17,14 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
     const query = collection(db, 'Usuarios');
     const [docs, loading, error] = useCollectionData(query);
     const [seriesIds, setSeriesIds] = useState([]);
-    const [currentPage, setCurrentPage] = useState(1);
+    const [currentPageMovies, setCurrentPageMovies] = useState(1);
+    const [currentPageSeries, setCurrentPageSeries] = useState(1);
     const ITEMS_PER_PAGE = 5;
     const navigate = useNavigate();
-    const mySwal = withReactContent(Swal);
+    //const mySwal = withReactContent(Swal);
 
+    const [showAlert, setShowAlert] = useState(false);  // Nuevo estado para controlar la visibilidad del Alert
+    const [alertConfig, setAlertConfig] = useState(null);
 
     const handleMovieIds = (ids, setDataFunction) => {
         if (movieIds.length !== ids.length) {
@@ -37,12 +39,13 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
     };
 
     //paginación
-    const indexOfLastItem = currentPage * ITEMS_PER_PAGE;
-    const indexOfFirstItem = indexOfLastItem - ITEMS_PER_PAGE;
-    const currentMovies = movies.slice(indexOfFirstItem, indexOfLastItem);
-    const currentSeries = series.slice(indexOfFirstItem, indexOfLastItem);
+    // const indexOfLastItem = currentPage * ITEMS_PER_PAGE;
+    // const indexOfFirstItem = indexOfLastItem - ITEMS_PER_PAGE;
+    // const currentMovies = movies.slice(indexOfFirstItem, indexOfLastItem);
+    // const currentSeries = series.slice(indexOfFirstItem, indexOfLastItem);
 
-    const paginate = (pageNumber) => setCurrentPage(pageNumber);
+    const paginateMovies = (pageNumber) => setCurrentPageMovies(pageNumber);
+    const paginateSeries = (pageNumber) => setCurrentPageSeries(pageNumber);
 
     const getData = async (ids, fetchDataFunction, setDataFunction) => {
         try {
@@ -59,7 +62,14 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
             getData(movieIds, getMovieById, setMovies);
             getData(seriesIds, getTVById, setSeries);
         }
-    }, [userEmail, movieIds, seriesIds]);
+
+        // Comprueba si se confirmó la eliminación
+        if (showAlert && alertConfig) {
+            // Actualiza el estado o realiza cualquier otra acción
+            setShowAlert(false); // Restablece el estado de showAlert
+            setAlertConfig(null); // Restablece el estado de alertConfig
+        }
+    }, [userEmail, movieIds, seriesIds, showAlert,  alertConfig]);
 
     const handleDelete = async (id, type, setDataFunction) => {
         try {
@@ -72,9 +82,25 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
         if (docToDelete) {
             await deleteDoc(docToDelete.ref);
             setDataFunction((prevItems) => prevItems.filter((item) => item.id !== id));
-            console.log('Item deleted successfully');
+            setShowAlert(true)
+            setAlertConfig({
+                title:"Eliminado!",
+                text:"El item ha sido eliminado.",
+                icon:"success",
+                showCancelButton: false,
+                confirmButtonText: "OK",
+            })
+            // console.log('Item deleted successfully');
             } else {
-                console.log('Document not found for item with ID:', id);
+                setShowAlert(true)
+                setAlertConfig({
+                    title:"Advertencia!",
+                    text:`Document not found for item with ID: ${id}`,
+                    icon:"warning",
+                    showCancelButton: false,
+                    confirmButtonText: "OK",
+                })
+                //console.log('Document not found for item with ID:', id);
             }
         } catch (error) {
             console.error('Error deleting item:', error);
@@ -82,29 +108,17 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
     };
 
     const confirmDelete = (id, type, setDataFunction) => {
-        mySwal
-            .fire({
-                title: "¿Estas seguro?",
-                text: "¡No podrás revertir esto!",
-                icon: "warning",
-                showCancelButton: true,
-                confirmButtonColor: "#d33",
-                cancelButtonColor: "#3085d6",
-                confirmButtonText: "¡Si, Eliminar!",
-                cancelButtonText: "Cancelar",
-            })
-            .then((result) => {
-                if (result.isConfirmed) {
-                // Llamo a la función para eliminar
-                    handleDelete(id, type, setDataFunction);
-                        mySwal.fire(
-                            "Eliminado!",
-                            "El item ha sido eliminado.",
-                            "success",
-                        );
-                    }
-            });
-        };
+        setShowAlert(true)
+        setAlertConfig({
+            title: "¿Estas seguro?",
+            text: "¡No podrás revertir esto!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "¡Si, Eliminar!",
+            cancelButtonText: "Cancelar",
+            onConfirm: () => handleDelete(id, type, setDataFunction)
+        })
+    };
 
     return (
         <>
@@ -117,28 +131,40 @@ const RenderItemsContainer = ({ userEmail, pathMovies, pathSeries }) => {
                 <div style={{width:'100%', display: 'flex', flexDirection: 'column'}}>
                     {movies.length > 0 && (
                         <RenderItems
-                            items={currentMovies}
+                            items={movies.slice((currentPageMovies - 1) * ITEMS_PER_PAGE, currentPageMovies * ITEMS_PER_PAGE)}
                             itemsPaginado={movies}
                             type='peliculas'
                             setItemsFunction={setMovies}
-                            handleDelete={confirmDelete}
+                            handleDelete={(id) => confirmDelete(id, 'peliculas', setMovies)}
                             navigate={navigate}
                             ITEMS_PER_PAGE={ITEMS_PER_PAGE}
-                            paginate={paginate}/>
+                            paginate={paginateMovies}/>
                     )}
 
                     {series.length > 0 && (
                         <RenderItems
-                            items={currentSeries}
+                            items={series.slice((currentPageSeries - 1) * ITEMS_PER_PAGE, currentPageSeries * ITEMS_PER_PAGE)}
                             itemsPaginado={series}
                             type='series'
                             setItemsFunction={setSeries}
-                            handleDelete={confirmDelete}
+                            handleDelete={(id) => confirmDelete(id, 'series', setSeries)}
                             navigate={navigate}
                             ITEMS_PER_PAGE={ITEMS_PER_PAGE}
-                            paginate={paginate}/>
+                            paginate={paginateSeries}/>
                     )}
                 </div>
+            )}
+            {/* Alerta para eliminar */}
+            {showAlert && alertConfig && (
+                <Alert
+                    title={alertConfig.title}
+                    text={alertConfig.text}
+                    icon={alertConfig.icon}
+                    showCancelButton={alertConfig.showCancelButton}
+                    confirmButtonText={alertConfig.confirmButtonText}
+                    cancelButtonText={alertConfig.cancelButtonText}
+                    onConfirm={alertConfig.onConfirm}
+                />
             )}
         </>
     );
