+++ src/context/authContext.js
@@ -15,9 +15,26 @@ export function AuthProvider({ children }) {
     const [user, setUser] = useState(null);
     const [loading, setLoading] = useState(true);
 
+    // Obtener el valor de 'showVideo' del almacenamiento local
+    const showVideoFromLocalStorage = localStorage.getItem("showVideo");
+    const [showVideo, setShowVideo] = useState(showVideoFromLocalStorage !== "false");
+
+    // useEffect para actualizar 'showVideo' cuando el usuario se deslogue
+    // useEffect(() => {
+    //     if (!user) {
+    //     // El usuario se ha deslogueado, actualiza 'showVideo' a true
+    //     setShowVideo(true);
+    //     }
+    // }, [user]);
+
     const signup = (email, password) => createUserWithEmailAndPassword(auth, email, password);
     const login = async (email, password) => signInWithEmailAndPassword(auth, email, password);
-    const logout = () => signOut(auth);
+    const logout = () => {
+        // Antes de cerrar sesión, actualiza 'showVideo' en el localStorage a true
+        localStorage.setItem("showVideo", "true");
+        setShowVideo(true);
+        signOut(auth);
+    }
     const loginWithGoogle = () => {
         const GoogleProvider = new GoogleAuthProvider();
         return signInWithPopup(auth, GoogleProvider);
@@ -38,6 +55,12 @@ export function AuthProvider({ children }) {
         return () => unsubscribe();
     }, []);
 
+    useEffect(() => {
+        // Almacenar el valor actual de 'showVideo' en el localStorage
+        localStorage.setItem("showVideo", showVideo.toString());
+    }, [showVideo]);
+
+
     const value = {
         isAuth,
         setIsAuth,
@@ -47,7 +70,9 @@ export function AuthProvider({ children }) {
         logout,
         loading,
         loginWithGoogle,
-        resetPassword
+        resetPassword,
+        showVideo, 
+        setShowVideo
     };
 
     return (
