+++ src/pages/Home/Home.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import LayoutMain from '../../layout/LayoutMain/LayoutMain'
 import {HomeDeskoptComponents} from './HomeDeskopt/HomeDeskopt';
 import HomeMobile  from './HomeMobile/HomeMobile';
-import { useAuth } from '../../context/authContext2';  // Importa el hook useAuth
+import { useAuth } from '../../context/authContext';  // Importa el hook useAuth
 import { useResponsive } from '../../context/responsiveContext.js'
 
 const Home = () => {
