+++ src/pages/Home/HomeDeskopt/HomeDeskopt.jsx
@@ -1,6 +1,6 @@
 import {ButtonGroup} from 'react-bootstrap';
 import Boton from '../../../components/Boton/Boton'
-import Banner from '../../../components/Banner/Banner2'
+import Banner from '../../../components/Banner/Banner'
 import CarruselVertical from '../../../components/Carrusel/CarruselVertical/CarruselVertical'
 import {useNavigate} from 'react-router'
 import './HomeDeskopt.css'
