+++ src/pages/LoginRegistro/LoginRegistro.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import LayoutForm from '../../layout/LayoutForm/LayoutForm';
-import { Login } from '../../components/Ingreso/Login/Login2';
+import { Login } from '../../components/Ingreso/Login/Login';
 import { Register } from '../../components/Ingreso/Register/Register';
 import Boton from '../../components/Boton/Boton';
-import './LoginRegistro.css';
 import colors from '../../config/config.js';
+import './LoginRegistro.css';
+
 
 const LoginRegistro = () => {
   return (
