+++ src/pages/Recientes/Recientes.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import LayoutSecundario from '../../layout/LayoutSecundario/LayoutSecundario';
-import RenderItemsContainer from '../../components/RenderItems/RenderItemsContainer/RenderItemsContainer2'
-import { useAuth } from '../../context/authContext2.js';
+import RenderItemsContainer from '../../components/RenderItems/RenderItemsContainer/RenderItemsContainer'
+import { useAuth } from '../../context/authContext.js';
 
 const Recientes = () => {
   const { user } = useAuth();
